BMTraits.cs removed
AgentExtensions.cs
- HasTrait (allows checking for vanilla traits with the vanilla enums)

BMTraitController.cs
 added more multi-trait-relevant methods

+++ Content/Traits/T_Tampering/TamperTantrum.cs
@@ -33,9 +33,9 @@ namespace BunnyMod.Content.Traits
 
 		public override void OnRemoved() { }
 
-		public static double GetToolCostFactor(Agent agent)
+		public static float GetToolCostFactor(Agent agent)
 		{
-			return agent.HasTrait<TamperTantrum>() ? 0.5d : 1d;
+			return agent.HasTrait<TamperTantrum>() ? 0.5f : 1f;
 		}
 		
 		public static void AgentInteractions_AddButton_Prefix(string buttonName, ref string extraCost, Agent mostRecentInteractingAgent)
